src/third_party/blink/web_tests/virtual/not-site-per-process/external/wpt/html/browsers/origin/relaxing-the-same-origin-restriction/document_domain_access_details.sub-expected.txt

This is a testharness.js-based test.
PASS Access allowed if same-origin with no 'document.domain' modification. (Sanity check)
PASS Access not allowed if different-origin with no 'document.domain' modification. (Sanity check)
PASS Access disallowed if same-origin but only one sets document.domain.
PASS Access allowed if same-origin and both set document.domain to existing value.
PASS Access allowed if different-origin but both set document.domain to parent domain.
PASS Access disallowed again if same-origin, both set document-domain to existing value, then one sets to parent.
FAIL Access is revoked to Window object when we stop being same effective script origin due to document.domain. assert_equals: expected "SecurityError" but got "Reachable 1"
PASS Access is revoked to Location object when we stop being same effective script origin due to document.domain.
PASS Access is not revoked to Document object when we stop being same effective script origin due to document.domain.
PASS Access is not revoked to random object when we stop being same effective script origin due to document.domain.
Harness: the test ran to completion.
